web/requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
Jinja2
openpyxl